test(auth): localize school role fixtures
