+++ backend/services/auth/auth_login_school_cutover_test.go
@@ -13,14 +13,53 @@ import (
 	"context"
 	"fmt"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	authModels "github.com/moto-nrw/project-phoenix/models/auth"
 	"github.com/moto-nrw/project-phoenix/services/auth"
 	testpkg "github.com/moto-nrw/project-phoenix/test"
 	"github.com/uptrace/bun"
 )
+
+func assignSchoolSystemRole(t *testing.T, db *bun.DB, accountID, tenantID int64, roleName string) {
+	t.Helper()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	var roleID int64
+	require.NoError(t, db.NewSelect().
+		ColumnExpr("id").
+		TableExpr("auth.roles").
+		Where("name = ?", roleName).
+		Where("is_system = TRUE").
+		Scan(ctx, &roleID), "seeded %s system role must exist", roleName)
+
+	assignment := &authModels.AccountRole{AccountID: accountID, RoleID: roleID}
+	assignment.SetTenantID(tenantID)
+	_, err := db.NewInsert().
+		Model(assignment).
+		ModelTableExpr(`auth.account_roles`).
+		Exec(ctx)
+	require.NoError(t, err, "assign seeded %s system role", roleName)
+}
+
+func removeSchoolSystemRole(t *testing.T, db *bun.DB, accountID, tenantID int64, roleName string) {
+	t.Helper()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	_, err := db.NewDelete().
+		TableExpr("auth.account_roles AS ar").
+		Where("ar.account_id = ?", accountID).
+		Where("ar.tenant_id = ?", tenantID).
+		Where("ar.role_id IN (SELECT id FROM auth.roles WHERE name = ? AND is_system = TRUE)", roleName).
+		Exec(ctx)
+	require.NoError(t, err, "remove seeded %s system role", roleName)
+}
 
 // registerAccountAtTenant creates an account mapped to a fresh tenant and
 // returns its credentials plus the tenant it belongs to.
@@ -80,7 +119,7 @@ func TestTenantLogin_DualRoleLehrkraftPassesThrough(t *testing.T) {
 	testpkg.AssignLehrkraftSystemRole(t, db, accountID, tenantID)
 	// Same person is also a Betreuungskraft at this school — small schools do
 	// exactly this, and the cutover must not lock them out of the OGS portal.
-	testpkg.AssignSystemRoleByName(t, db, accountID, tenantID, "user")
+	assignSchoolSystemRole(t, db, accountID, tenantID, "user")
 
 	accessToken, refreshToken, err := service.LoginWithAudit(context.Background(), email, testPassword, "", "", slug)
 
@@ -112,7 +151,7 @@ func TestRefreshToken_SchoolPortalOnlyAccountIsRefused(t *testing.T) {
 
 	email, accountID, tenantID, slug := registerAccountAtTenant(t, db, service, "refresh-school-only")
 	testpkg.AssignLehrkraftSystemRole(t, db, accountID, tenantID)
-	testpkg.AssignSystemRoleByName(t, db, accountID, tenantID, "user")
+	assignSchoolSystemRole(t, db, accountID, tenantID, "user")
 
 	_, refreshToken, err := service.LoginWithAudit(context.Background(), email, testPassword, "", "", slug)
 	require.NoError(t, err)
@@ -120,7 +159,7 @@ func TestRefreshToken_SchoolPortalOnlyAccountIsRefused(t *testing.T) {
 	// The caregiver role is withdrawn while the session is still open. A
 	// tenant session that keeps renewing itself would outlive the very role
 	// that justified it.
-	testpkg.RemoveSystemRoleByName(t, db, accountID, tenantID, "user")
+	removeSchoolSystemRole(t, db, accountID, tenantID, "user")
 
 	_, _, err = service.RefreshTokenWithAudit(context.Background(), refreshToken, "", "")
